+++ account/client.go
@@ -39,7 +39,7 @@ import (
 	"github.com/ontio/ontology/common/password"
 	"github.com/ontio/ontology/core/types"
 	ontErrors "github.com/ontio/ontology/errors"
-	"github.com/ontio/ontology/net/protocol"
+	"github.com/ontio/ontology/p2pserver/protocol"
 )
 
 const (
