+++ p2pserver/actor/net_server.go
@@ -21,9 +21,9 @@ package actor
 import (
 	"reflect"
 
-	"github.com/ontio/ontology/common/log"
-	"github.com/ontio/ontology/net/protocol"
 	"github.com/ontio/ontology-eventbus/actor"
+	"github.com/ontio/ontology/common/log"
+	"github.com/ontio/ontology/p2pserver/protocol"
 )
 
 var netServerPid *actor.PID
