+++ p2pserver/message/address.go
@@ -28,7 +28,7 @@ import (
 	"strconv"
 
 	"github.com/ontio/ontology/common/log"
-	"github.com/ontio/ontology/net/protocol"
+	"github.com/ontio/ontology/p2pserver/protocol"
 )
 
 type addrReq struct {
