+++ p2pserver/message/not_found.go
@@ -22,78 +22,85 @@ import (
 	"bytes"
 	"crypto/sha256"
 	"encoding/binary"
+	"errors"
 
+	"github.com/ontio/ontology/common"
 	"github.com/ontio/ontology/common/log"
-	"github.com/ontio/ontology/common/serialization"
-	"github.com/ontio/ontology/net/actor"
-	"github.com/ontio/ontology/net/protocol"
+	"github.com/ontio/ontology/p2pserver/protocol"
 )
 
-type pong struct {
+type notFound struct {
 	msgHdr
-	height uint64
+	hash common.Uint256
 }
 
-func NewPongMsg() ([]byte, error) {
-	var msg pong
+func NewNotFound(hash common.Uint256) ([]byte, error) {
+	log.Debug()
+	var msg notFound
+	msg.hash = hash
 	msg.msgHdr.Magic = protocol.NET_MAGIC
-	copy(msg.msgHdr.CMD[0:7], "pong")
-	height, _ := actor.GetCurrentHeaderHeight()
-	msg.height = uint64(height)
+	cmd := "notfound"
+	copy(msg.msgHdr.CMD[0:len(cmd)], cmd)
 	tmpBuffer := bytes.NewBuffer([]byte{})
-	serialization.WriteUint64(tmpBuffer, msg.height)
-	b := new(bytes.Buffer)
-	err := binary.Write(b, binary.LittleEndian, tmpBuffer.Bytes())
+	msg.hash.Serialize(tmpBuffer)
+	p := new(bytes.Buffer)
+	err := binary.Write(p, binary.LittleEndian, tmpBuffer.Bytes())
 	if err != nil {
-		log.Error("Binary Write failed at new Msg")
+		log.Error("Binary Write failed at new notfound Msg")
 		return nil, err
 	}
-	s := sha256.Sum256(b.Bytes())
+	s := sha256.Sum256(p.Bytes())
 	s2 := s[:]
 	s = sha256.Sum256(s2)
 	buf := bytes.NewBuffer(s[:protocol.CHECKSUM_LEN])
 	binary.Read(buf, binary.LittleEndian, &(msg.msgHdr.Checksum))
-	msg.msgHdr.Length = uint32(len(b.Bytes()))
+	msg.msgHdr.Length = uint32(len(p.Bytes()))
+	log.Debug("The message payload length is ", msg.msgHdr.Length)
 
 	m, err := msg.Serialization()
 	if err != nil {
 		log.Error("Error Convert net message ", err.Error())
 		return nil, err
 	}
+
 	return m, nil
 }
 
-func (msg pong) Verify(buf []byte) error {
+func (msg notFound) Verify(buf []byte) error {
 	err := msg.msgHdr.Verify(buf)
 	return err
 }
 
-func (msg pong) Handle(node protocol.Noder) error {
-	node.SetHeight(msg.height)
-	return nil
-}
-
-func (msg pong) Serialization() ([]byte, error) {
+func (msg notFound) Serialization() ([]byte, error) {
 	hdrBuf, err := msg.msgHdr.Serialization()
 	if err != nil {
 		return nil, err
 	}
 	buf := bytes.NewBuffer(hdrBuf)
-	err = serialization.WriteUint64(buf, msg.height)
-	if err != nil {
-		return nil, err
-	}
-	return buf.Bytes(), err
+	msg.hash.Serialize(buf)
 
+	return buf.Bytes(), err
 }
 
-func (msg *pong) Deserialization(p []byte) error {
+func (msg *notFound) Deserialization(p []byte) error {
 	buf := bytes.NewBuffer(p)
+
 	err := binary.Read(buf, binary.LittleEndian, &(msg.msgHdr))
 	if err != nil {
-		return err
+		log.Warn("Parse notfound message hdr error")
+		return errors.New("Parse notfound message hdr error")
 	}
 
-	msg.height, err = serialization.ReadUint64(buf)
+	err = msg.hash.Deserialize(buf)
+	if err != nil {
+		log.Warn("Parse notfound message error")
+		return errors.New("Parse notfound message error")
+	}
+
 	return err
+}
+
+func (msg notFound) Handle(node protocol.Noder) error {
+	log.Debug("RX notfound message, hash is ", msg.hash)
+	return nil
 }
